Add more stdout

+++ matlab/kill-template.sh
@@ -12,9 +12,13 @@ fi
 bash ${remote_session_dir}/kill-vnc-${job_number}.sh
 
 service_pid=$(cat ${remote_session_dir}/kill.pid)
-echo "Killing process $(hostname):${service_pid}"
-pkill -P ${service_pid}
-kill ${service_pid}
+if [ -z ${chdir} ]; then
+    echo "ERROR: No service pid was found!"
+else
+    echo "Killing process $(hostname):${service_pid}"
+    pkill -P ${service_pid}
+    kill ${service_pid}
+fi
 
 #kill $(ps -x | grep MATLAB | awk '{print $1}')
 #kill $(ps -x | grep matlab | awk '{print $1}')
